+++ sockeye/checkpoint_decoder.py
@@ -133,13 +133,17 @@ class CheckpointDecoder:
             [checkpoint],
             softmax_temperature=self.softmax_temperature,
             max_output_length_num_stds=self.max_output_length_num_stds)
-        translator = inference.Translator(self.context,
-                                          self.ensemble_mode,
-                                          self.bucket_width_source,
-                                          inference.LengthPenalty(self.length_penalty_alpha, self.length_penalty_beta),
-                                          models,
-                                          source_vocabs,
-                                          target_vocab)
+        translator = inference.Translator(context=self.context,
+                                          ensemble_mode=self.ensemble_mode,
+                                          bucket_source_width=self.bucket_width_source,
+                                          length_penalty=inference.LengthPenalty(self.length_penalty_alpha, self.length_penalty_beta),
+                                          beam_prune=0.0,
+                                          beam_search_stop='all',
+                                          models=models,
+                                          source_vocabs=source_vocabs,
+                                          target_vocab=target_vocab,
+                                          restrict_lexicon=None,
+                                          store_beam=False)
         trans_wall_time = 0.0
         translations = []
         with data_io.smart_open(output_name, 'w') as output: